Added android support

+++ android/app/src/main/java/dev/fanchao/CJKProxy.java
@@ -3,8 +3,8 @@ package dev.fanchao;
 import androidx.annotation.NonNull;
 
 public class CJKProxy {
-    public native static long start(@NonNull String upstreamHost, short upstreamPort,
-                                    @NonNull String socks5Host, short socks5Port) throws Exception;
+    public native static long start(@NonNull String upstreamHost, int upstreamPort,
+                                    @NonNull String socks5Host, int socks5Port) throws Exception;
 
     public native static void stop(long instance);
